--- chat/src/docs/chatrooms.adoc
@@ -1,33 +1,0 @@
-
-== Chatting
-=== 텃밭 채팅방 만들기
-텃밭 분양 채팅방을 만들 수 있다.
-
-==== Request
-include::{snippets}/create-garden-chat-room/http-request.adoc[]
-include::{snippets}/create-garden-chat-room/request-fields.adoc[]
-
-==== Response (Success)
-include::{snippets}/create-garden-chat-room/http-response.adoc[]
-
-=== 작물거래 채팅방 만들기
-작물 거래 채팅방을 만들 수 있다.
-
-==== Request
-include::{snippets}/create-crop-chat-room/http-request.adoc[]
-include::{snippets}/create-crop-chat-room/request-fields.adoc[]
-
-==== Response (Success)
-include::{snippets}/create-crop-chat-room/http-response.adoc[]
-
-
-=== 텃밭 채팅방 입장
-텃밭 채팅방 입장할 때는 읽지 않은 상대방의 메세지는 모두 읽음 처리로 변하며 +
-텃밭 채팅방 상단 탭에 노출되는 정보들을 응답으로 반환한다.
-
-==== Request
-include::{snippets}/enter-garden-chat-room/http-request.adoc[]
-include::{snippets}/enter-garden-chat-room/path-parameters.adoc[]
-
-==== Response (Success)
-include::{snippets}/enter-garden-chat-room/http-response.adoc[]
