--- chat/src/main/java/com/garden/back/auth/AuthProvider.java
@@ -1,5 +1,0 @@
-package com.garden.back.auth;
-
-public enum AuthProvider {
-    KAKAO, NAVER
-}
