--- chat/src/main/java/com/garden/back/auth/client/MemberProvider.java
@@ -1,7 +1,0 @@
-package com.garden.back.auth.client;
-
-import com.garden.back.member.Member;
-
-public interface MemberProvider {
-    Member getMember(String authorization);
-}
