--- chat/src/main/java/com/garden/back/auth/client/naver/NaverMemberProvider.java
@@ -1,22 +1,0 @@
-package com.garden.back.auth.client.naver;
-
-import com.garden.back.auth.client.naver.response.NaverOauth2Response;
-import com.garden.back.member.Member;
-import com.garden.back.auth.client.MemberProvider;
-import org.springframework.stereotype.Component;
-
-@Component
-public class NaverMemberProvider implements MemberProvider {
-
-    private final NaverOAuthClient naverClient;
-
-    public NaverMemberProvider(NaverOAuthClient naverClient) {
-        this.naverClient = naverClient;
-    }
-
-    @Override
-    public Member getMember(String authorization) {
-        NaverOauth2Response response = naverClient.getUserInfoFromKakao(authorization);
-        return response.toEntity();
-    }
-}
