--- chat/src/main/java/com/garden/back/auth/jwt/repository/RefreshToken.java
@@ -1,11 +1,0 @@
-package com.garden.back.auth.jwt.repository;
-
-import com.garden.back.member.Member;
-
-import java.util.Date;
-
-public record RefreshToken(
-    String key,
-    Member member,
-    Date expiredDate
-) {}
