--- chat/src/main/java/com/garden/back/auth/jwt/repository/RefreshTokenRepository.java
@@ -1,11 +1,0 @@
-package com.garden.back.auth.jwt.repository;
-
-import java.util.Optional;
-
-public interface RefreshTokenRepository<T> {
-    void save(T refreshToken);
-
-    void deleteExpiredToken();
-
-    Optional<T> findByKey(String key);
-}
