--- chat/src/main/java/com/garden/back/auth/jwt/response/RefreshTokenResponse.java
@@ -1,6 +1,0 @@
-package com.garden.back.auth.jwt.response;
-
-public record RefreshTokenResponse(
-    String accessToken
-) {
-}
