--- chat/src/main/java/com/garden/back/auth/jwt/response/TokenResponse.java
@@ -1,10 +1,0 @@
-package com.garden.back.auth.jwt.response;
-
-public record TokenResponse(
-    String grantType,
-    String accessToken,
-    String refreshToken,
-    Long accessTokenExpiredDate,
-    Long refreshTokenExpiredDate
-) {
-}
