--- chat/src/main/java/com/garden/back/facade/dto/GardenChatRoomEnterFacadeRequest.java
@@ -1,27 +1,0 @@
-package com.garden.back.facade.dto;
-
-import com.garden.back.global.loginuser.LoginUser;
-import com.garden.back.service.dto.request.ChatRoomEntryParam;
-
-public record GardenChatRoomEnterFacadeRequest(
-        Long roomId,
-        Long memberId
-) {
-    public ChatRoomEntryParam to() {
-        return new ChatRoomEntryParam(
-                roomId,
-                memberId
-        );
-    }
-
-    public static GardenChatRoomEnterFacadeRequest to(
-            Long roomId,
-            LoginUser loginUser
-    ) {
-        return new GardenChatRoomEnterFacadeRequest(
-                roomId,
-                loginUser.memberId()
-        );
-    }
-
-}
