--- chat/src/main/java/com/garden/back/global/validation/EnumValue.java
@@ -1,17 +1,0 @@
-package com.garden.back.global.validation;
-
-import jakarta.validation.Constraint;
-import jakarta.validation.Payload;
-
-import java.lang.annotation.*;
-
-@Target({ElementType.FIELD, ElementType.PARAMETER, ElementType.LOCAL_VARIABLE})
-@Retention(RetentionPolicy.RUNTIME)
-@Documented
-@Constraint(validatedBy = EnumValueValidator.class)
-public @interface EnumValue {
-    Class<? extends Enum> enumClass();
-    String message() default "유효하지 않은 Enum 입니다.";
-    Class<?>[] groups() default {};
-    Class<? extends Payload>[] payload() default {};
-}
