--- chat/src/test/java/com/garden/back/docs/RestDocsSupport.java
@@ -1,36 +1,0 @@
-package com.garden.back.docs;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.data.web.PageableHandlerMethodArgumentResolver;
-import org.springframework.restdocs.RestDocumentationContextProvider;
-import org.springframework.restdocs.RestDocumentationExtension;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import org.springframework.web.filter.CharacterEncodingFilter;
-
-import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.documentationConfiguration;
-import static org.springframework.restdocs.operation.preprocess.Preprocessors.prettyPrint;
-
-@ExtendWith(RestDocumentationExtension.class)
-public abstract class RestDocsSupport {
-
-    protected MockMvc mockMvc;
-
-    protected ObjectMapper objectMapper = new ObjectMapper();
-
-    @BeforeEach
-    void setUp(RestDocumentationContextProvider provider) {
-        this.mockMvc = MockMvcBuilders.standaloneSetup(initController())
-                .setCustomArgumentResolvers(new PageableHandlerMethodArgumentResolver())
-                .addFilter(new CharacterEncodingFilter("UTF-8", true))
-                .apply(documentationConfiguration(provider)
-                        .operationPreprocessors()
-                        .withRequestDefaults(prettyPrint())
-                        .withResponseDefaults(prettyPrint()))
-                .build();
-    }
-
-    protected abstract Object initController();
-}
